Anchors update - Anchor shading added

+++ main.cpp
@@ -5,6 +5,6 @@
 
 int main() {
 	ieml::Node node{ieml::file("../exampleSrc/test.ieml")};
-	std::cout << node["first"][0].as<int>() << "\n";
+	std::cout << node["first"][0].as<double>() << "\n";
 	std::cout << node;
 }
